relocate plot related helper functions
