+++ neuralprophet/forecaster.py
@@ -14,6 +14,7 @@ from neuralprophet import configure, df_utils, metrics, time_dataset, time_net, 
 from neuralprophet.logger import MetricsLogger
 from neuralprophet.plot_forecast import plot, plot_components
 from neuralprophet.plot_forecast_plotly import plot as plot_plotly
+from neuralprophet.plot_forecast_plotly import get_valid_configuration
 from neuralprophet.plot_forecast_plotly import plot_components as plot_components_plotly
 from neuralprophet.plot_model_parameters import plot_parameters
 from neuralprophet.plot_model_parameters_plotly import plot_parameters as plot_parameters_plotly
@@ -1740,413 +1741,6 @@ class NeuralProphet:
                 line_per_origin=True,
             )
 
-    def check_if_configured(self, components, error_flag=False):
-        """Check if components were set in the model configuration by the user.
-
-        Parameters
-        ----------
-            components : str or list, optional
-                name or list of names of components to check
-
-                Options
-                ----
-                * ``trend``
-                * ``trend_rate_change``
-                * ``seasonality``
-                * ``autoregression``
-                * ``lagged_regressors```
-                * ``events``
-                * ``future_regressors`
-                * ``residuals``
-                * ``uncertainty``
-            error_flag : bool
-                Activate to raise a ValueError if component has not been configured
-
-        Returns
-        -------
-            components
-                list of components only including the components set in the model configuration
-        """
-        invalid_components = []
-        if "trend_rate_change" in components and self.model.config_trend.changepoints is None:
-            components.remove("trend_rate_change")
-            invalid_components.append("trend_rate_change")
-        if "seasonality" in components and self.config_season is None:
-            components.remove("seasonality")
-            invalid_components.append("seasonality")
-        if "autoregression" in components and not self.config_ar.n_lags > 0:
-            components.remove("autoregression")
-            invalid_components.append("autoregression")
-        if "lagged_regressors" in components and self.config_lagged_regressors is None:
-            components.remove("lagged_regressors")
-            invalid_components.append("lagged_regressors")
-        if "events" in components and (self.config_events and self.config_country_holidays) is None:
-            components.remove("events")
-            invalid_components.append("events")
-        if "future_regressors" in components and self.config_regressors is None:
-            components.remove("future_regressors")
-            invalid_components.append("future_regressors")
-        if "uncertainty" in components and not len(self.model.quantiles) > 1:
-            components.remove("uncertainty")
-            invalid_components.append("uncertainty")
-        if error_flag and len(invalid_components) != 0:
-            raise ValueError(
-                f" Selected component(s) {(invalid_components)} for plotting not specified in the model configuration."
-            )
-        return components
-
-    def get_valid_configuration(
-        self, components=None, df_name=None, valid_set=None, validator=None, forecast_in_focus=None, quantile=0.5
-    ):
-        """Validate and adapt the selected components to be plotted.
-
-        Parameters
-        ----------
-            components : str or list, optional
-                name or list of names of components to validate and adapt
-            df_name: str
-                ID from time series that should be plotted
-            valid_set : str or list, optional
-                name or list of names of components that are defined as valid option
-
-                Options
-                ----
-                 * (default)``None``:  All components the user set in the model configuration are validated and adapted
-                * ``trend``
-                * ``seasonality``
-                * ``autoregression``
-                * ``lagged_regressors``
-                * ``future_regressors``
-                * ``events``
-                * ``residuals``
-                * ``uncertainty``
-            validator: str
-                specifies the validation purpose to customize
-
-                Options
-                ----
-                * ``plot_parameters``: customize for plot_parameters() function
-                * ``plot_components``: customize for plot_components() function
-            forecast_in_focus: int
-                optinal, i-th step ahead forecast to plot
-
-                Note
-                ----
-                None (default): plot self.highlight_forecast_step_n by default
-            quantile: float
-                The quantile for which the model parameters are to be plotted
-
-                Note
-                ----
-                0.5 (default):  Parameters will be plotted for the median quantile.
-
-        Returns
-        -------
-            valid_configuration: dict
-                dict of validated components and values to be plotted
-        """
-        if type(valid_set) is not list:
-            valid_set = [valid_set]
-
-        if components is None:
-            components = valid_set
-            components = self.check_if_configured(components)
-        else:
-            if type(components) is not list:
-                components = [components]
-            components = [comp.lower() for comp in components]
-            for comp in components:
-                if comp not in valid_set:
-                    raise ValueError(
-                        f" Selected component {comp} is either mis-spelled or not an available "
-                        f"option for this function."
-                    )
-            components = self.check_if_configured(components, error_flag=True)
-        if validator is None:
-            raise ValueError("Specify a validator from the available options")
-        # Adapt Normalization
-        if validator == "plot_parameters":
-            # Set to True in case of local normalization and unknown_data_params is not True
-            overwriting_unknown_data_normalization = False
-            if self.config_normalization.global_normalization:
-                if df_name is None:
-                    df_name = "__df__"
-                else:
-                    log.debug("Global normalization set - ignoring given df_name for normalization")
-            else:
-                if df_name is None:
-                    log.warning("Local normalization set, but df_name is None. Using global data params instead.")
-                    df_name = "__df__"
-                    if not self.config_normalization.unknown_data_normalization:
-                        self.config_normalization.unknown_data_normalization = True
-                        overwriting_unknown_data_normalization = True
-                elif df_name not in self.config_normalization.local_data_params:
-                    log.warning(
-                        f"Local normalization set, but df_name '{df_name}' not found. Using global data params instead."
-                    )
-                    df_name = "__df__"
-                    if not self.config_normalization.unknown_data_normalization:
-                        self.config_normalization.unknown_data_normalization = True
-                        overwriting_unknown_data_normalization = True
-                else:
-                    log.debug(f"Local normalization set. Data params for {df_name} will be used to denormalize.")
-
-        # Identify components to be plotted
-        # as dict, minimum: {plot_name}
-        plot_components = []
-        if validator == "plot_parameters":
-            quantile_index = self.model.quantiles.index(quantile)
-
-        # Plot trend
-        if "trend" in components:
-            plot_components.append({"plot_name": "Trend", "comp_name": "trend"})
-        if "trend_rate_change" in components:
-            plot_components.append({"plot_name": "Trend Rate Change"})
-
-        # Plot  seasonalities, if present
-        if "seasonality" in components:
-            for name in self.config_season.periods:
-                if validator == "plot_components":
-                    plot_components.append(
-                        {
-                            "plot_name": f"{name} seasonality",
-                            "comp_name": name,
-                        }
-                    )
-                elif validator == "plot_parameters":
-                    plot_components.append({"plot_name": "seasonality", "comp_name": name})
-
-        # AR
-        if "autoregression" in components:
-            if validator == "plot_components":
-                if forecast_in_focus is None:
-                    plot_components.append(
-                        {
-                            "plot_name": "Auto-Regression",
-                            "comp_name": "ar",
-                            "num_overplot": self.n_forecasts,
-                            "bar": True,
-                        }
-                    )
-                else:
-                    plot_components.append(
-                        {
-                            "plot_name": f"AR ({forecast_in_focus})-ahead",
-                            "comp_name": f"ar{forecast_in_focus}",
-                        }
-                    )
-            elif validator == "plot_parameters":
-                plot_components.append(
-                    {
-                        "plot_name": "lagged weights",
-                        "comp_name": "AR",
-                        "weights": self.model.ar_weights.detach().numpy(),
-                        "focus": forecast_in_focus,
-                    }
-                )
-
-        # Add lagged regressors
-        lagged_scalar_regressors = []
-        if "lagged_regressors" in components:
-            if validator == "plot_components":
-                if forecast_in_focus is None:
-                    for name in self.config_lagged_regressors.keys():
-                        plot_components.append(
-                            {
-                                "plot_name": f'Lagged Regressor "{name}"',
-                                "comp_name": f"lagged_regressor_{name}",
-                                "num_overplot": self.n_forecasts,
-                                "bar": True,
-                            }
-                        )
-                else:
-                    for name in self.config_lagged_regressors.keys():
-                        plot_components.append(
-                            {
-                                "plot_name": f'Lagged Regressor "{name}" ({forecast_in_focus})-ahead',
-                                "comp_name": f"lagged_regressor_{name}{forecast_in_focus}",
-                            }
-                        )
-            elif validator == "plot_parameters":
-                for name in self.config_lagged_regressors.keys():
-                    if self.config_lagged_regressors[name].as_scalar:
-                        lagged_scalar_regressors.append((name, self.model.get_covar_weights(name).detach().numpy()))
-                    else:
-                        plot_components.append(
-                            {
-                                "plot_name": "lagged weights",
-                                "comp_name": f'Lagged Regressor "{name}"',
-                                "weights": self.model.get_covar_weights(name).detach().numpy(),
-                                "focus": forecast_in_focus,
-                            }
-                        )
-
-        # Add Events
-        additive_events = []
-        multiplicative_events = []
-        if "events" in components:
-            additive_events_flag = False
-            muliplicative_events_flag = False
-            for event, configs in self.config_events.items():
-                if validator == "plot_components" and configs.mode == "additive":
-                    additive_events_flag = True
-                elif validator == "plot_components" and configs.mode == "multiplicative":
-                    muliplicative_events_flag = True
-                elif validator == "plot_parameters":
-                    event_params = self.model.get_event_weights(event)
-                    weight_list = [
-                        (key, param.detach().numpy()[quantile_index, :]) for key, param in event_params.items()
-                    ]
-                    if configs.mode == "additive":
-                        additive_events = additive_events + weight_list
-                    elif configs.mode == "multiplicative":
-                        multiplicative_events = multiplicative_events + weight_list
-
-            for country_holiday in self.config_country_holidays.holiday_names:
-                if validator == "plot_components" and self.config_country_holidays.mode == "additive":
-                    additive_events_flag = True
-                elif validator == "plot_components" and self.config_country_holidays.mode == "multiplicative":
-                    muliplicative_events_flag = True
-                elif validator == "plot_parameters":
-                    event_params = self.model.get_event_weights(country_holiday)
-                    weight_list = [
-                        (key, param.detach().numpy()[quantile_index, :]) for key, param in event_params.items()
-                    ]
-                    if self.config_country_holidays.mode == "additive":
-                        additive_events = additive_events + weight_list
-                    elif self.config_country_holidays.mode == "multiplicative":
-                        multiplicative_events = multiplicative_events + weight_list
-
-            if additive_events_flag:
-                plot_components.append(
-                    {
-                        "plot_name": "Additive Events",
-                        "comp_name": "events_additive",
-                    }
-                )
-            if muliplicative_events_flag:
-                plot_components.append(
-                    {
-                        "plot_name": "Multiplicative Events",
-                        "comp_name": "events_multiplicative",
-                        "multiplicative": True,
-                    }
-                )
-
-        # Add Regressors
-        additive_future_regressors = []
-        multiplicative_future_regressors = []
-        if "future_regressors" in components:
-            for regressor, configs in self.config_regressors.items():
-                if validator == "plot_components" and configs.mode == "additive":
-                    plot_components.append(
-                        {
-                            "plot_name": "Additive Future Regressors",
-                            "comp_name": "future_regressors_additive",
-                        }
-                    )
-                elif validator == "plot_components" and configs.mode == "multiplicative":
-                    plot_components.append(
-                        {
-                            "plot_name": "Multiplicative Future Regressors",
-                            "comp_name": "future_regressors_multiplicative",
-                            "multiplicative": True,
-                        }
-                    )
-                elif validator == "plot_parameters":
-                    regressor_param = self.model.get_reg_weights(regressor)[quantile_index, :]
-                    if configs.mode == "additive":
-                        additive_future_regressors.append((regressor, regressor_param.detach().numpy()))
-                    elif configs.mode == "multiplicative":
-                        multiplicative_future_regressors.append((regressor, regressor_param.detach().numpy()))
-
-        if "residuals" in components and validator == "plot_components":
-            if forecast_in_focus is None and self.n_forecasts > 1:
-                plot_components.append(
-                    {
-                        "plot_name": "Residuals",
-                        "comp_name": "residual",
-                        "num_overplot": self.n_forecasts,
-                        "bar": True,
-                    }
-                )
-            else:
-                ahead = 1 if forecast_in_focus is None else forecast_in_focus
-                plot_components.append(
-                    {
-                        "plot_name": f"Residuals ({ahead})-ahead",
-                        "comp_name": f"residual{ahead}",
-                        "bar": True,
-                    }
-                )
-        # Plot  quantiles as a separate component, if present
-        # If multiple steps in the future are predicted, only plot quantiles if highlight_forecast_step_n is set
-        if (
-            "quantiles" in components
-            and validator == "plot_components"
-            and len(self.model.quantiles) > 1
-            and forecast_in_focus is None
-        ):
-            if len(self.config_train.quantiles) > 1 and (
-                self.n_forecasts > 1 or self.config_ar.n_lags > 0
-            ):  # rather query if n_forecasts >1 than n_lags>1
-                raise ValueError(
-                    "Please specify step_number using the highlight_nth_step_ahead_of_each_forecast function"
-                    " for quantiles plotting when autoregression enabled."
-                )
-            for i in range(1, len(self.model.quantiles)):
-                plot_components.append(
-                    {
-                        "plot_name": "Uncertainty",
-                        "comp_name": f"yhat1 {round(self.model.quantiles[i] * 100, 1)}%",
-                        "fill": True,
-                    }
-                )
-        elif (
-            "uncertainty" in components
-            and validator == "plot_components"
-            and len(self.model.quantiles) > 1
-            and forecast_in_focus is not None
-        ):
-            for i in range(1, len(self.model.quantiles)):
-                plot_components.append(
-                    {
-                        "plot_name": "Uncertainty",
-                        "comp_name": f"yhat{forecast_in_focus} {round(self.model.quantiles[i] * 100, 1)}%",
-                        "fill": True,
-                    }
-                )
-        if validator == "plot_parameters":
-            if len(additive_future_regressors) > 0:
-                plot_components.append({"plot_name": "Additive future regressor"})
-            if len(multiplicative_future_regressors) > 0:
-                plot_components.append({"plot_name": "Multiplicative future regressor"})
-            if len(lagged_scalar_regressors) > 0:
-                plot_components.append({"plot_name": "Lagged scalar regressor"})
-            if len(additive_events) > 0:
-                data_params = self.config_normalization.get_data_params(df_name)
-                scale = data_params["y"].scale
-                additive_events = [(key, weight * scale) for (key, weight) in additive_events]
-                plot_components.append({"plot_name": "Additive event"})
-            if len(multiplicative_events) > 0:
-                plot_components.append({"plot_name": "Multiplicative event"})
-
-            valid_configuration = {
-                "components_list": plot_components,
-                "additive_future_regressors": additive_future_regressors,
-                "additive_events": additive_events,
-                "multiplicative_future_regressors": multiplicative_future_regressors,
-                "multiplicative_events": multiplicative_events,
-                "lagged_scalar_regressors": lagged_scalar_regressors,
-                "overwriting_unknown_data_normalization": overwriting_unknown_data_normalization,
-                "df_name": df_name,
-            }
-        elif validator == "plot_components":
-            valid_configuration = {
-                "components_list": plot_components,
-            }
-        return valid_configuration
-
     def plot_components(
         self,
         fcst,
@@ -2249,7 +1843,8 @@ class NeuralProphet:
             "residuals",
             "uncertainty",
         ]
-        valid_plot_configuration = self.get_valid_configuration(
+        valid_plot_configuration = get_valid_configuration(
+            m=self,
             components=components,
             df_name=df_name,
             valid_set=valid_components_set,
@@ -2372,7 +1967,7 @@ class NeuralProphet:
         # Check if highlighted forecast step is overwritten
         if forecast_in_focus is None:
             forecast_in_focus = self.highlight_forecast_step_n
-        if (self.highlight_forecast_step_n or forecast_in_focus) is not None and self.self.config_ar.n_lags == 0:
+        if (self.highlight_forecast_step_n or forecast_in_focus) is not None and self.config_ar.n_lags == 0:
             log.warning("highlight_forecast_step_n is ignored since autoregression not enabled.")
             forecast_in_focus = None
         if forecast_in_focus is not None and forecast_in_focus > self.n_forecasts:
@@ -2408,7 +2003,8 @@ class NeuralProphet:
             "events",
             "future_regressors",
         ]
-        valid_plot_configuration = self.get_valid_configuration(
+        valid_plot_configuration = get_valid_configuration(
+            m=self,
             components=components,
             df_name=df_name,
             forecast_in_focus=forecast_in_focus,
